requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
google-auth
google-api-python-client
pydantic
python-dotenv
